+++ lib/notion/connection.ts
@@ -1,4 +1,11 @@
-import { Client as NotionClient } from "@notionhq/client";
+import {
+  DatabaseObjectResponse,
+  Client as NotionClient,
+  PageObjectResponse
+} from "@notionhq/client";
+import { QueryDatabaseParameters } from "@notionhq/client/build/src/api-endpoints";
+import { logger } from "@sentry/nextjs";
+import { DATABASES } from "./types";
 
 const NOTION_TOKEN = process.env.NOTION_API_KEY;
 
@@ -13,3 +20,87 @@ export default function getConnection() {
     throw error;
   }
 }
+
+/**
+ * Retrieves the Notion database ID for the specified database type.
+ * Throws an error if the database type is not defined or if the corresponding
+ * environment variable for the database ID is missing.
+ *
+ * @param database - The type of database to retrieve the ID for (e.g., 'cases' or 'stack').
+ * @returns The Notion database ID as a string.
+ * @throws Will throw an error if the database type is not provided or if the database ID is not defined in environment variables.
+ */
+
+export function getDatabaseIdOrError(database: DATABASES): string {
+  if (!database) {
+    logger.fatal("Error fetching database", { error: "Database type is not defined" });
+    throw new Error("Database type is not defined");
+  }
+
+  const databasesByName = {
+    cases: process.env.NOTION_CASES_DATABASE_ID,
+    stack: process.env.NOTION_STACK_DATABASE_ID
+  };
+  const NOTION_DATABASE_ID = databasesByName[database] || null;
+
+  if (!NOTION_DATABASE_ID) {
+    logger.fatal("Error fetching database", { error: "NOTION_DATABASE_ID is not defined" });
+    throw new Error("NOTION_DATABASE_ID is not defined");
+  }
+  return NOTION_DATABASE_ID;
+}
+
+/**
+ * Retrieves the properties of a Notion database.
+ * @param database The database to query.
+ * @returns The database object response.
+ */
+export const getDatabase = async (database: DATABASES): Promise<DatabaseObjectResponse> => {
+  try {
+    const databaseId = getDatabaseIdOrError(database);
+    logger.trace("Getting database", { databaseId, database });
+    const notion = getConnection();
+    const retrievedDB = await notion.databases.retrieve({ database_id: databaseId });
+    return retrievedDB as DatabaseObjectResponse;
+  } catch (error) {
+    logger.fatal("Error fetching database", { error, data: { database } });
+    throw error;
+  }
+};
+
+/**
+ * Queries all entries (pages) from a Notion database, handling pagination.
+ * @param database The database to query.
+ * @param queryOptions Optional query parameters like filters and sorts.
+ * @returns An array of page objects representing the database entries.
+ */
+export const queryDatabase = async (
+  database: DATABASES,
+  queryOptions?: Partial<QueryDatabaseParameters>
+): Promise<PageObjectResponse[]> => {
+  try {
+    logger.trace("Starting database connection", { database });
+    const notion = getConnection();
+    const databaseId = getDatabaseIdOrError(database);
+    let allEntries: PageObjectResponse[] = [];
+    let cursor: string | null = null;
+
+    do {
+      const response = await notion.databases.query({
+        database_id: databaseId,
+        page_size: 100,
+        start_cursor: cursor || undefined,
+        ...queryOptions
+      });
+
+      const pages = response.results as PageObjectResponse[];
+      allEntries = allEntries.concat(pages);
+      cursor = response.next_cursor;
+    } while (cursor);
+
+    return allEntries;
+  } catch (error) {
+    logger.fatal("Error querying database with ID", { database, error });
+    throw error;
+  }
+};
